Refatoração: -Remoção da classe Descrição -Inclusão das classes ViewHolder dentro das classes Adapter
