+++ app/src/main/java/com/example/pokedex/adapter/PokemonDialogAdapter.kt
@@ -2,28 +2,59 @@ package com.example.pokedex.adapter
 
 import android.content.Context
 import android.view.LayoutInflater
+import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.pokedex.R
+import com.example.pokedex.controller.MudaCor
 import com.example.pokedex.model.Pokemon
-import com.example.pokedex.viewholder.DialogViewHolder
+import kotlinx.android.synthetic.main.row_dialog_fragment.view.*
 
-class PokemonDialogAdapter(pokemon: Pokemon, appContext : Context) : RecyclerView.Adapter<DialogViewHolder>() {
+class PokemonDialogAdapter(private val pokemon: Pokemon) : RecyclerView.Adapter<PokemonDialogAdapter.PokemonDialogViewHolder>() {
 
-    var pokemonSelecionado = pokemon
-    val appContext = appContext
+    lateinit var context : Context
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): DialogViewHolder {
-        val inflater = LayoutInflater.from(parent.context)
-        val view = inflater.inflate(R.layout.row_dialog_fragment, parent, false)
-        return DialogViewHolder(view, appContext)
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PokemonDialogViewHolder {
+        context = parent.context
+
+        val itemView = LayoutInflater.from(parent.context).inflate(R.layout.row_dialog_fragment, parent, false)
+        return PokemonDialogViewHolder(itemView, context)
     }
 
-    override fun onBindViewHolder(holder: DialogViewHolder, position: Int) {
-        holder.bindInfo(pokemonSelecionado)
+    override fun onBindViewHolder(holder: PokemonDialogViewHolder, position: Int) {
+        holder.bindView(pokemon)
     }
 
-    override fun getItemCount(): Int {
-        return 1
+    override fun getItemCount(): Int = 1
+
+    class PokemonDialogViewHolder(private val view : View,
+                                  private val context : Context) : RecyclerView.ViewHolder(view){
+        fun bindView(pokemon : Pokemon){
+            val imagemPokemon = view.imagemPokemon
+            val textNumeroPokemon = view.textNumeroPokemon
+            val textNomePokemon = view.textNomePokemon
+            val textDescricao = view.textDescricao
+            val textTipo1 = view.textTipo1
+            val textTipo2 = view.textTipo2
+            val uriImagem = "p${pokemon.numero}"
+
+            imagemPokemon.setImageResource(caminhoImagem(context, uriImagem))
+            textNumeroPokemon.text = pokemon.numero.toString()
+            textNomePokemon.text = pokemon.nome
+            textDescricao.text = pokemon.descricao
+            textTipo1.text = pokemon.tipo1
+            MudaCor().mudarCorTipo(textTipo1, context)
+
+            if (textTipo2.text != null) {
+                textTipo2.text = pokemon.tipo2
+                MudaCor().mudarCorTipo(textTipo2, view.context)
+            }
+        }
+
+        //Retorna um caminho a partir de uma String
+        fun caminhoImagem(context: Context, nomeImagem: String): Int {
+            return context.resources
+                .getIdentifier("drawable/$nomeImagem", null, context.packageName)
+        }
     }
 }
