--- app/src/main/java/com/example/pokedex/model/Description.kt
@@ -1,262 +1,0 @@
-package com.example.pokedex.model
-
-enum class Description(val descricao: String) {
-
-    //Primeira geração
-
-    BULBASAUR("Pokémon Semente. \n\nBulbasaur pode ser visto cochilando sob a luz do sol. Há uma semente nas costas. Ao absorver os raios do sol, a semente cresce progressivamente."),
-    IVYSAUR("Pokémon Semente. \n\nHá um broto nas costas deste Pokémon. Para suportar seu peso, as pernas e o tronco de Ivysaur ficam grossos e fortes. Se começar a passar mais tempo deitado sob a luz do sol, é um sinal de que o broto florescerá em uma grande flor em breve."),
-    VENUSAUR("Pokémon Semente. \n\nHá uma flor grande nas costas de Venusaur. Diz-se que a flor adquire cores vivas se receber muita nutrição e luz solar. O aroma da flor acalma as emoções das pessoas."),
-    CHARMANDER("Pokémon Lagarto. \n\nA chama que queima na ponta da cauda é uma indicação de suas emoções. A chama tremula quando Charmander está se divertindo. Se o Pokémon ficar enfurecido, a chama queima ferozmente."),
-    CHARMELEON("Pokémon Chama. \n\nCharmeleon impiedosamente destrói seus inimigos usando suas garras afiadas. Se encontrar um inimigo forte, ele se torna agressivo. Nesse estado eufórico, a chama na ponta de sua cauda brilha com uma cor branca azulada."),
-    CHARIZARD("Pokémon Chama. \n\nCharizard voa pelo céu em busca de oponentes poderosos. Respira um fogo tão quente que derrete qualquer coisa. No entanto, nunca usa seu sopro em um oponente mais fraco que ele."),
-    SQUIRTLE("Pokémon Tartaruga. \n\nA concha de Squirtle não é usada apenas para proteção. A forma arredondada da concha e as ranhuras na superfície ajudam a minimizar a resistência na água, permitindo que este Pokémon nade em alta velocidade."),
-    WARTORTLE("Pokémon Tartaruga. \n\nSua cauda é grande e coberta com um pêlo rico e grosso. A cor de sua cauda fica mais escura à medida que Wartortle envelhece. Os arranhões em sua concha são uma evidência da dureza deste Pokémon como um lutador."),
-    BLASTOISE("Pokémon Crustáceo. \n\nBlastoise tem canhões de água que se projetam de sua concha. Os canhões de água são muito precisos. Eles podem disparar balas de água com precisão suficiente para atingir latas vazias a uma distância de mais de 60 metros."),
-    CATERPIE("Pokémon Minhoca. \n\nCaterpie tem um apetite voraz. Pode devorar folhas maiores do que seu corpo diante de seus olhos. Da sua antena, este Pokémon libera um odor terrivelmente forte."),
-    METAPOD("Pokémon Casulo. \n\nA concha que cobre o corpo deste Pokémon é tão dura quanto uma placa de ferro. Metapod não se move muito. Ele fica parado porque está preparando suas entranhas suaves para a evolução dentro da casca dura."),
-    BUTTERFREE("Pokémon Borboleta. \n\nButterfree tem uma capacidade superior de procurar néctar delicioso de flores. Pode até buscar, extrair e transportar néctar de flores que estão desabrochando a mais de 10 quilômetros do ninho."),
-    WEEDLE("Pokémon Inseto Peludo. \n\nWeedle tem um olfato extremamente agudo. É capaz de distinguir seus tipos favoritos de folhas daqueles de que não gosta, apenas farejando com seu grande nariz vermelho."),
-    KAKUNA("Pokémon Casulo. \n\nKakuna permanece praticamente imóvel quando se agarra a uma árvore. No entanto, por dentro, é extremamente ocupado, enquanto se prepara para a sua evolução futura. Isso é evidente pelo quão quente a concha fica ao toque."),
-    BEEDRILL("Pokémon Abelha Venenosa. \n\nBeedrill é extremamente territorial. Ninguém jamais deve se aproximar de seu ninho - isso é para sua própria segurança. Se irritados, eles atacarão em um enxame furioso."),
-    PIDGEY("Pokémon Passarinho. \n\nPidgey tem um senso de direção extremamente afiado. Ela é capaz de voltar para o seu ninho de forma infalível, por mais longe que possa ser removida de seu ambiente familiar."),
-    PIDGEOTTO("Pokémon Pássaro. \n\nPidgeotto reivindica uma grande área como seu próprio território. Este Pokémon voa, patrulhando o espaço em que vive. Se seu território é violado, não mostra piedade de punir completamente o inimigo com suas garras afiadas."),
-    PIDGEOT("Pokémon Pássaro. \n\nEste Pokémon tem uma plumagem deslumbrante de penas maravilhosamente brilhantes. Muitos treinadores são cativados pela impressionante beleza das penas em sua cabeça, obrigando-os a escolher Pidgeot como seu Pokémon."),
-    RATTATA("Pokémon Camundongo. \n\nRattata é cauteloso ao extremo. Mesmo dormindo, ouve constantemente movendo os ouvidos. Não é exigente sobre onde vive - fará seu ninho em qualquer lugar."),
-    RATICATE("Pokémon Camundongo. \n\nAs presas fortes de Raticate crescem constantemente. Para mantê-los triturados, roem pedras e troncos. Pode até roer as paredes das casas."),
-    SPEAROW("Pokémon Passarinho. \n\nSpearow tem um grito muito alto que pode ser ouvido a mais de 800 metros de distância. Se seu grito alto e agudo for ouvido ecoando por toda parte, é um sinal de que eles estão alertando um ao outro sobre o perigo."),
-    FEAROW("Pokémon Bicudo. \n\nFearow é reconhecido pelo pescoço longo e pelo bico alongado. Eles são convenientemente modelados para capturar presas no solo ou na água. Ele habilmente move seu bico longo e magro para colher presas."),
-    EKANS("Pokémon Serpente. \n\nEkans se enrola em espiral enquanto descansa. Assumir esta posição permite que ela responda rapidamente a uma ameaça de qualquer direção com um olhar penetrante de sua cabeça erguida."),
-    ARBOK("Pokémon Cobra. \n\nEste Pokémon é terrivelmente forte para apertar as coisas com seu corpo. Pode até amassar tonéis de aço. Uma vez que Arbok envolve seu corpo em torno de seu inimigo, é impossível escapar do abraço esmagador."),
-    PIKACHU("Pokémon Rato. \n\nSempre que Pikachu se depara com algo novo, ele o acerta com um choque elétrico. Se você se deparar com uma fruta queimada, é evidencia de que este Pokémon exagerou na intensidade de sua carga."),
-    RAICHU("Pokémon Rato. \n\nSe os sacos elétricos ficam excessivamente carregados, Raichu planta sua cauda no chão e descarrega. Pedaços de terra queimados serão encontrados perto do ninho deste Pokémon."),
-    SANDSHREW("Pokémon Rato. \n\nO corpo de Sandshrew está configurado para absorver água sem desperdício, permitindo que ele sobreviva em um deserto árido. Este Pokémon se enrola para se proteger de seus inimigos."),
-    SANDSLASH("Pokémon Rato. \n\nO corpo do Sandslash é coberto por pontas duras, que são seções endurecidas de sua pele. Uma vez por ano, os espinhos antigos caem, para serem substituídos por novos espinhos que crescem por baixo deles."),
-    NIDORANF("Pokémon Espinho Venenoso. \n\nNidoran♀ tem farpas que secretam um poderoso veneno. Pensa-se que eles tenham se desenvolvido como proteção para este Pokémon de corpo pequeno. Quando enfurecido, libera uma toxina horrível do chifre."),
-    NIDORINA("Pokémon Espinho Venenoso. \n\nQuando Nidorina está com seus amigos ou familiares, eles mantêm suas farpas ocultas para evitar se machucar. Este Pokémon parece ficar nervoso se separado dos outros."),
-    NIDOQUEEN("Pokémon Broca. \n\nO corpo de Nidoqueen é envolto em escamas extremamente duras. É hábil em arremessar inimigos com investidas severas. Este Pokémon é mais forte quando defende seus filhotes."),
-    NIDORANM("Pokémon Espinho Venenoso. \n\nNidoran♂ desenvolveu músculos para mover seus ouvidos. Graças a eles, os ouvidos podem ser movidos livremente em qualquer direção. Mesmo o menor som não escapa à atenção deste Pokémon."),
-    NIDORINO("Pokémon Espinho Venenoso. \n\nNidorino tem um chifre mais duro que um diamante. Se sente uma presença hostil, todas as farpas nas costas se arrepiam de uma só vez, e ele desafia o inimigo com toda a sua força."),
-    NIDOKING("Pokémon Broca. \n\nA cauda grossa de Nidoking possui enorme poder destrutivo. Com um balanço, ele pode derrubar uma torre de transmissão de metal. Uma vez que este Pokémon entra em fúria, não há como detê-lo."),
-    CLEFAIRY("Pokémon Fada. \n\nEm todas as noites de lua cheia, grupos deste Pokémon aparecem para brincar. Quando amanhece, os Clefairy cansados voltam para seus retiros tranquilos nas montanhas e vão dormir aninhados."),
-    CLEFABLE("Pokémon Fada. \n\nClefable se move pulando levemente como se estivesse voando usando as asas. Seu passo saltitante permite que ele ande sobre a água. É conhecido por passear em lagos em noites tranquilas e iluminadas pela lua."),
-    VULPIX("Pokémon Raposa. \n\nNo momento do seu nascimento, o Vulpix tinha uma cauda branca. A cauda se separa em seis se este Pokémon receber muito amor de seu treinador. As seis caudas ficam magnificamente enroladas."),
-    NINETALES("Pokémon Raposa. \n\nNinetales lança uma luz sinistra de seus olhos vermelhos brilhantes para obter controle total sobre a mente de seu inimigo. Diz-se que este Pokémon vive por mil anos."),
-    JIGGLYPUFF("Pokémon Balão. \n\nAs cordas vocais de Jigglypuff podem ajustar livremente o comprimento de onda de sua voz. Este Pokémon usa essa capacidade de cantar exatamente no comprimento de onda certo para deixar seus inimigos mais sonolentos."),
-    WIGGLYTUFF("Pokémon Balão. \n\nWigglytuff tem olhos grandes e parecidos com pires. As superfícies dos olhos estão sempre cobertas por uma fina camada de lágrimas. Se houver poeira nos olhos deste Pokémon, ela é rapidamente lavada."),
-    ZUBAT("Pokémon Morcego. \n\nZubat permanece silenciosamente imóvel em um ponto escuro durante o dia claro. Isso ocorre porque a exposição prolongada ao sol faz com que seu corpo se torne ligeiramente queimado."),
-    GOLBAT("Pokémon Morcego. \n\nGolbat adora beber o sangue de seres vivos. É particularmente ativo no escuro da noite. Este Pokémon voa pelo céu noturno, buscando sangue fresco."),
-    ODDISH("Pokémon Erva. \n\nDurante o dia, Oddish se enterra no solo para absorver os nutrientes do solo usando todo o corpo. Quanto mais fértil o solo, mais brilhantes ficam as suas folhas."),
-    GLOOM("Pokémon Erva. \n\nGloom libera uma fragrância fétida do pistilo de sua flor. Quando confrontado com o perigo, o fedor piora. Se este Pokémon estiver calmo e seguro, ele não liberará seu aroma fedido habitual."),
-    VILEPLUME("Pokémon Flor. \n\nO pólen tóxico de Vileplume desencadeia ataques alérgicos atrozes. É por isso que é aconselhável nunca abordar flores atraentes em uma floresta, por mais bonitas que sejam."),
-    PARAS("Pokémon Cogumelo. \n\nParas tem cogumelos parasitas crescendo nas costas, chamados tochukaso. Eles crescem ao extrair nutrientes deste Pokémon Inseto hospedeiro. Eles são altamente valorizados como remédio para prolongar a vida."),
-    PARASECT("Pokémon Cogumelo.\n\nParasect é conhecido por infestar grandes árvores em massa e drenar nutrientes do tronco e raízes mais baixas. Quando uma árvore infestada morre, elas se movem para outra árvore de uma só vez."),
-    VENONAT("Pokémon Inseto.\n\nDiz-se que Venonat evoluiu com uma camada de cabelo fino e rígido que cobre todo o corpo para proteção. Possui olhos grandes que nunca falham em detectar presas minúsculas."),
-    VENOMOTH("Pokémon Mariposa Venenosa.\n\nVenomoth é noturno - é um Pokémon que só se torna ativo à noite. Suas presas favoritas são pequenos insetos que se reúnem em torno das luzes da rua, atraídos pela luz na escuridão."),
-    DIGLETT("Pokémon Toupeira.\n\nDiglett são criados na maioria das fazendas. O motivo é simples: onde quer que esse Pokémon se enterre, o solo fica perfeitamente cultivado para o plantio. Este solo é ideal para o cultivo de vegetais deliciosos."),
-    DUGTRIO("Pokémon Toupeira.\n\nDugtrio são na verdade trigêmeos que emergiram de um corpo. Como resultado, cada trigêmeo pensa exatamente como os outros dois trigêmeos. Eles trabalham cooperativamente para escavar sem parar."),
-    MEOWTH("Pokémon Arranha Gato.\n\nMeowth retira suas garras afiadas em suas patas para esgueirar-se lentamente sem dar passos incriminadores. Por alguma razão, este Pokémon adora moedas brilhantes que brilham com luz."),
-    PERSIAN("Pokémon Gato Elegante.\n\nPersian tem seis bigodes grossos que lhe dão um ar de durão. Os bigodes percebem os movimentos do ar para determinar o que há nas proximidades do Pokémon. Torna-se dócil se agarrado pelos bigodes."),
-    PSYDUCK("Pokémon Pato.\n\nPsyduck usa um poder misterioso. Quando isso acontece, este Pokémon gera ondas cerebrais que supostamente são vistas apenas em pessoas que estão dormindo. Essa descoberta gerou controvérsias entre os estudiosos."),
-    GOLDUCK("Pokémon Pato.\n\nAs nadadeiras com membranas nas patas dianteiras e traseiras e o corpo aerodinâmico de Golduck proporcionam uma velocidade assustadora. Este Pokémon é definitivamente muito mais rápido do que o nadador mais atlético."),
-    MANKEY("Pokémon Macaco Porco.\n\nQuando Mankey começa a tremer e sua respiração nasal fica áspera, é um sinal claro de que está ficando com raiva. No entanto, como entra em uma fúria iminente quase instantaneamente, é impossível alguém fugir de sua ira."),
-    PRIMEAPE("Pokémon Macaco Porco.\n\nQuando Primeape fica furioso, sua circulação sanguínea é aumentada. Por sua vez, seus músculos são fortalecidos. No entanto, também se torna muito menos inteligente ao mesmo tempo."),
-    GROWLITHE("Pokémon Filhote.\n\nGrowlithe tem um excelente olfato. Uma vez que cheira algo, este Pokémon não esquece o aroma, não importa o quê. Ele usa seu sentido olfativo avançado para determinar as emoções de outros seres vivos."),
-    ARCANINE("Pokémon Lendário.\n\nArcanine é conhecido por sua alta velocidade. Diz-se que é capaz de percorrer mais de 10.000 quilômetros em um único dia e noite. O fogo que arde violentamente dentro do corpo deste Pokémon é sua fonte de energia."),
-    POLIWAG("Pokémon Girino.\n\nPoliwag tem uma pele muito fina. É possível ver as entranhas em espiral do Pokémon através da pele. Apesar da sua magreza, a pele também é muito flexível. Até presas afiadas ricocheteiam."),
-    POLIWHIRL("Pokémon Girino.\n\nA superfície do corpo de Poliwhirl está sempre molhada e escorregadia com um fluido viscoso. Por causa dessa cobertura escorregadia, ele pode facilmente escorregar e deslizar para fora das garras de qualquer inimigo em batalha."),
-    POLIWRATH("Pokémon .\n\nOs músculos fortes e desenvolvidos de Poliwrath nunca se cansam, por mais que exercite. É tão incansavelmente forte que este Pokémon pode nadar para frente e para trás através do oceano sem esforço."),
-    ABRA("Pokémon Psi.\n\nAbra dorme dezoito horas por dia. No entanto, ele pode sentir a presença de inimigos enquanto está dormindo. Em tal situação, este Pokémon se teleporta imediatamente para a segurança."),
-    KADABRA("Pokémon Psi.\n\nKadabra emite uma onda alfa peculiar se desenvolver dor de cabeça. Somente as pessoas com uma psique particularmente forte podem esperar se tornar um treinador deste Pokémon."),
-    ALAKAZAM("Pokémon Psi.\n\nO cérebro de Alakazam cresce continuamente, tornando a cabeça pesada demais para suportar o pescoço. Este Pokémon mantém sua cabeça erguida usando seu poder psicocinético."),
-    MACHOP("Pokémon Superpoder.\n\nOs músculos de Machop são especiais - eles nunca ficam doloridos, por mais que sejam usados em exercícios. Este Pokémon tem poder suficiente para arremessar uma centena de humanos adultos."),
-    MACHOKE("Pokémon Superpoder.\n\nOs músculos completamente tonificados de Machoke possuem a dureza do aço. Este Pokémon tem tanta força que pode segurar no alto um lutador de sumô com apenas um dedo."),
-    MACHAMP("Pokémon Superpoder.\n\nMachamp tem o poder de arremessar qualquer coisa. No entanto, tentar fazer qualquer trabalho que exija cuidado e destreza faz com que seus braços se enrosquem. Este Pokémon tende a entrar em ação antes de pensar."),
-    BELLSPROUT("Pokémon Flor.\n\nO corpo fino e flexível do Bellsprout permite que ele se curve e balance para evitar qualquer ataque, por mais forte que seja. Da sua boca, este Pokémon cospe um fluido corrosivo que derrete até ferro."),
-    WEEPINBELL("Pokémon Papa-Moscas.\n\nWeepinbell tem um gancho grande na extremidade traseira. À noite, o Pokémon se conecta a um galho de árvore e vai dormir. Se ele se move durante o sono, pode acordar e se encontrar no chão."),
-    VICTREEBELL("Pokémon Papa-Moscas.\n\nVictreebel tem uma longa videira que se estende de sua cabeça. Esta videira é balançada e sacudida como se fosse um animal para atrair presas. Quando uma presa desavisada se aproxima, esse Pokémon o engole inteiro."),
-    TENTACOOL("Pokémon Água Viva.\n\nO corpo do Tentacool é basicamente composto de água. Se for removido do mar, seca como um pergaminho. Se este Pokémon ficar desidratado, coloque-o novamente no mar."),
-    TENTACRUEL("Pokémon Água Viva.\n\nTentacruel tem grandes esferas vermelhas em sua cabeça. Os orbes brilham antes de atacar em volta com uma forte explosão ultrassônica. A explosão deste Pokémon cria ondas violentas ao seu redor."),
-    GEODUDE("Pokémon Pedra.\n\nQuanto mais um Geodude vive, mais suas bordas são lascadas e desgastadas, tornando-o mais arredondada na aparência. No entanto, o coração deste Pokémon permanecerá duro e áspero sempre."),
-    GRAVELER("Pokémon Pedra.\n\nGraveler cresce alimentando-se de rochas. Aparentemente, prefere comer pedras cobertas de musgo. Este Pokémon come uma tonelada de rochas diariamente."),
-    GOLEM("Pokémon Megaton.\n\nGolem vive nas montanhas. Se houver um grande terremoto, esses Pokémon virão rolando das montanhas em massa até o sopé abaixo."),
-    PONYTA("Pokémon Cavalo de Fogo.\n\nPonyta é muito fraco ao nascer. Ele mal consegue se levantar. Este Pokémon se torna mais forte tropeçando e caindo para acompanhar seus pais."),
-    RAPIDASH("Pokémon Cavalo de Fogo.\n\nRapidash geralmente pode ser visto casualmente galopando nos campos e planícies. No entanto, quando esse Pokémon se torna sério, suas crinas ardentes brilham e queimam enquanto galopam até 150 mph."),
-    SLOWPOKE("Pokémon Bobo.\n\nSlowpoke usa sua cauda para capturar presas mergulhando-a na água ao lado de um rio. No entanto, esse Pokémon geralmente esquece o que está fazendo e muitas vezes passa dias inteiros apenas vagando à beira da água."),
-    SLOWBRO("Pokémon Caranguejo Eremita.\n\nA cauda de Slowbro tem um Shellder firmemente preso com uma mordida. Como resultado, a cauda não pode mais ser usada para pescar. Isso faz com que o Slowbro nade de má vontade e pegue presas."),
-    MAGNEMITE("Pokémon Imã.\n\nMagnemite se liga aos fios de energia para se alimentar de eletricidade. Se sua casa tiver uma queda de energia, verifique seus disjuntores. Você pode encontrar um grande número deste Pokémon agarrado à caixa do disjuntor."),
-    MAGNETON("Pokémon Imã.\n\nMagneton emite uma força magnética poderosa que é fatal para dispositivos mecânicos. Como resultado, grandes cidades soam sirenes para alertar os cidadãos sobre surtos em larga escala deste Pokémon."),
-    FARFETCHD("Pokémon Pato Selvagem.\n\nFarfetch'd é sempre visto com um caule de uma planta de algum tipo. Aparentemente, existem caules bons e caules ruins. Este Pokémon é conhecido por lutar com outros por caules."),
-    DODUO("Pokémon Pássaro Gêmeo.\n\nAs duas cabeças de Doduo nunca dormem ao mesmo tempo. Suas duas cabeças se revezam no sono, para que uma sempre possa vigiar os inimigos enquanto a outra dorme."),
-    DODRIO("Pokémon Pássaro Triplo.\n\nCuidado se as três cabeças de Dodrio estiverem olhando em três direções separadas. É um sinal claro de que está em guarda. Não chegue perto deste Pokémon se ele estiver cauteloso - ele pode decidir bicar você."),
-    SEEL("Pokémon Leão Marinho.\n\nSeel caça presas no mar gelado sob as camadas de gelo. Quando precisa respirar, faz um buraco no gelo com a parte acentuadamente saliente da cabeça."),
-    DEWGONG("Pokémon Leão Marinho.\n\nDewgong adora adormecer no gelo muito frio. A visão deste Pokémon dormindo em uma geleira foi confundida com uma sereia por um marinheiro há muito tempo."),
-    GRIMER("Pokémon Lodo.\n\nO corpo lamacento e emborrachado de Grimer pode ser forçado através de qualquer abertura, por menor que seja. Este Pokémon entra em canos de esgoto para beber água suja."),
-    MUK("Pokémon Lodo.\n\nDo corpo de Muk escoa um fluido fétido que emite um fedor horrível. Apenas uma gota do fluido corporal deste Pokémon pode tornar uma piscina estagnada e rançosa."),
-    SHELLDER("Pokémon Bivalve.\n\nÀ noite, esse Pokémon usa sua língua larga para cavar um buraco na areia do fundo do mar e depois dormir nela. Enquanto dorme, Shellder fecha a concha, mas deixa a língua para fora."),
-    CLOYSTER("Pokémon Bivalve.\n\nCloyster é capaz de nadar no mar. Isso é feito engolindo água e, em seguida, lançando-a para trás. Este Pokémon dispara espinhos de sua concha usando o mesmo sistema."),
-    GASTLY("Pokémon Gás.\n\nGastly é amplamente composto de matéria gasosa. Quando exposto a um vento forte, o corpo gasoso diminui rapidamente. Grupos deste Pokémon agrupam-se sob os beirais das casas para escapar da devastação do vento."),
-    HAUNTER("Pokémon Gás.\n\nHaunter é um Pokémon perigoso. Se um acena enquanto flutua na escuridão, nunca deve se aproximar dela. Este Pokémon tentará lamber você com a língua e roubar sua vida."),
-    GENGAR("Pokémon Sombra.\n\nÀs vezes, em uma noite escura, sua sombra lançada por uma luz da rua repentina e surpreendentemente o ultrapassa. Na verdade, é um Gengar passando por você, fingindo ser sua sombra."),
-    ONIX("Pokémon Serpente de Pedra.\n\nOnix tem um ímã em seu cérebro. Ele atua como uma bússola para que este Pokémon não perca a direção enquanto está escavando. À medida que envelhece, seu corpo se torna cada vez mais redondo e liso."),
-    DROWZEE("Pokémon Hipnose.\n\nSe seu nariz ficar com coceira enquanto você dorme, é um sinal claro de que um desses Pokémon está em cima de seu travesseiro e tenta comer seu sonho pelas narinas."),
-    HYPNO("Pokémon Hipnose.\n\nHypno segura um pêndulo na mão. O movimento de arco e o brilho do pêndulo embalam o inimigo em um profundo estado de hipnose. Enquanto este Pokémon procura presas, ele pole o pêndulo."),
-    KRABBY("Pokémon Caranguejo de Rio.\n\nKrabby vive nas praias, enterrados dentro de buracos cavados na areia. Em praias arenosas com pouco alimento, esses Pokémon podem ser vistos brigando entre si por território."),
-    KINGLER("Pokémon Pinça.\n\nKingler tem uma garra enorme. Ele agita essa enorme garra no ar para se comunicar com os outros. No entanto, como a garra é muito pesada, o Pokémon se cansa rapidamente."),
-    VOLTORB("Pokémon Bola.\n\nVoltorb foi avistado pela primeira vez em uma empresa que fabrica Pokébolas. A ligação entre esse avistamento e o fato de esse Pokémon parecer muito com uma Pokébola permanece um mistério."),
-    ELECTRODE("Pokémon Bola.\n\nElectrode consome eletricidade na atmosfera. Nos dias em que um raio atinge, você pode ver este Pokémon explodindo em todo o lugar por comer muita eletricidade."),
-    EXEGGCUTE("Pokémon Ovo.\n\nEste Pokémon consiste em seis ovos que formam um aglomerado. Os seis ovos se atraem e giram. Quando as rachaduras aparecem cada vez mais nos ovos, o Exeggcute está próximo da evolução."),
-    EXEGGUTOR("Pokémon Coqueiro.\n\nExeggutor veio originalmente dos trópicos. Suas cabeças crescem cada vez mais com a exposição à luz solar forte. Dizem que quando as cabeças caem, elas se agrupam para formar Exeggcute."),
-    CUBONE("Pokémon Solitário.\n\nCubone sofre pela mãe que nunca mais verá. Vendo a semelhança de sua mãe na lua cheia, ela chora. As manchas no crânio que o Pokémon usa são feitas pelas lágrimas que derrama."),
-    MAROWAK("Pokémon Guardião do Osso.\n\nMarowak é a forma evoluída de um Cubone que superou sua tristeza pela perda de sua mãe e se tornou durão. O espírito temperamental e duro deste Pokémon não é facilmente quebrado."),
-    HITMONLEE("Pokémon Chute.\n\nAs pernas de Hitmonlee se contraem e se esticam livremente. Usando essas pernas de mola, ele se curva sobre os inimigos com chutes devastadores. Após a batalha, esfrega as pernas e relaxa os músculos para superar a fadiga."),
-    HITMONCHAN("Pokémon Soco.\n\nDiz-se que Hitmonchan possui o espírito de um boxeador que estava treinando para um campeonato mundial. Este Pokémon tem um espírito indomável e nunca desistirá diante das adversidades."),
-    LICKITUNG("Pokémon Lambida.\n\nSempre que o Lickitung se deparar com algo novo, ele irá infalivelmente dar uma lambida. Faz isso porque memoriza as coisas por textura e por gosto. Não gosta muito de coisas azedas."),
-    KOFFING("Pokémon Gás Venenoso.\n\nSe Koffing fica agitado, aumenta a toxicidade de seus gases internos e os expele de todo o corpo. Este Pokémon também pode inflar demais o corpo redondo e explodir."),
-    WEEZING("Pokémon Gás Venenoso.\n\nWeezing ama os gases emitidos pelo lixo podre da cozinha. Este Pokémon encontrará uma casa suja e mal cuidada e a tornará sua casa. À noite, quando as pessoas da casa estão dormindo, ele passa pelo lixo."),
-    RHYHORN("Pokémon Espinhos.\n\nRhyhorn corre em linha reta, esmagando tudo em seu caminho. Não é atrapalhado, mesmo que passe por um bloco de aço. Este Pokémon pode sentir alguma dor da colisão no dia seguinte, no entanto."),
-    RHYDON("Pokémon Broca.\n\nO chifre de Rhydon pode esmagar até diamantes brutos. Um golpe arrebatador de sua cauda pode derrubar um edifício. A pele deste Pokémon é extremamente dura. Mesmo golpes diretos de canhão não deixam arranhões."),
-    CHANSEY("Pokémon Ovo.\n\nChansey põe ovos nutricionalmente excelentes todos os dias. Os ovos são tão deliciosos que são devorados com facilidade e avidez, mesmo pelas pessoas que perderam o apetite."),
-    TANGELA("Pokémon Videira.\n\nAs videiras de Tangela se soltam facilmente se forem agarradas. Isso acontece sem dor, permitindo que ele faça uma fuga rápida. As trepadeiras perdidas são substituídas por trepadeiras recém cultivadas no dia seguinte."),
-    KANGASKHAN("Pokémon Pai/Mãe.\n\nSe você se deparar com um jovem Kangaskhan brincando sozinho, nunca deve perturbá-lo ou tentar pegá-lo. O pai do bebê Pokémon certamente estará na área e ele ficará violentamente furioso com você."),
-    HORSEA("Pokémon Dragão.\n\nHorsea come pequenos insetos e musgo das rochas. Se a corrente oceânica gira rapidamente, este Pokémon se ancora envolvendo a cauda em torno de rochas ou corais para evitar ser arrastado."),
-    SEADRA("Pokémon Dragão.\n\nSeadra dorme depois de se contorcer entre corais. Aqueles que tentam coletar corais são ocasionalmente picados pelas farpas envenenadas deste Pokémon, se não perceberem."),
-    GOLDEEN("Pokémon Peixe Dourado.\n\nGoldeen é um Pokémon muito bonito, com barbatanas que ondulam elegantemente na água. No entanto, não abaixe a guarda em torno deste Pokémon - ele pode atacá-lo com força com seu chifre."),
-    SEAKING("Pokémon Peixe Dourado.\n\nNo outono, os Seaking machos podem ser vistos realizando danças reais nos leitos dos rios para atrair fêmeas. Durante esta temporada, a coloração corporal deste Pokémon é a mais bonita."),
-    STARYU("Pokémon Forma de Estrela.\n\nA parte central de Staryu tem um órgão chamado núcleo que brilha em vermelho brilhante. Se você for a uma praia no final do verão, os núcleos brilhantes desses Pokémon se parecem com as estrelas no céu."),
-    STARMIE("Pokémon Misterioso.\n\nA parte central de Starmie - o núcleo - brilha intensamente em sete cores. Devido à sua natureza luminosa, este Pokémon recebeu o apelido de \"a jóia do mar\"."),
-    MRMIME("Pokémon Barreira.\n\nMr. Mime é um mestre da pantomima. Seus gestos e movimentos convencem os observadores de que algo invisível realmente existe. Uma vez convencidos os observadores, o invisível existe como se fosse real."),
-    SCYTHER("Pokémon Louva a Deus.\n\nScyther é incrivelmente rápido. Sua velocidade incrível aumenta a eficácia das foices gêmeas em seus antebraços. As foices deste Pokémon são tão eficazes que podem cortar troncos grossos de uma só vez."),
-    JYNX("Pokémon Forma Humana.\n\nJynx caminha ritmicamente, balançando e rebolando os quadris como se estivesse dançando. Seus movimentos são tão atraentes que as pessoas são obrigadas a sacudir os quadris sem pensar no que estão fazendo."),
-    ELECTABUZZ("Pokémon Elétrico.\n\nQuando uma tempestade chega, as gangues deste Pokémon competem entre si para escalar alturas que provavelmente serão atingidas por raios. Algumas cidades usam o Electabuzz no lugar de pára-raios."),
-    MAGMAR("Pokémon Cospe Fogo.\n\nNa batalha, Magmar cospe chamas intensamente quentes de todo o corpo para intimidar seu oponente. As explosões ardentes deste Pokémon criam ondas de calor que incendeiam grama e árvores em seus arredores."),
-    PINSIR("Pokémon Besouro.\n\nPinsir é surpreendentemente forte. Ele pode segurar um inimigo pesando o dobro do seu peso nos chifres e levantá-lo facilmente. Os movimentos deste Pokémon ficam lentos em lugares frios."),
-    TAUROS("Pokémon Touro Selvagem.\n\nEste Pokémon não está satisfeito, a menos que esteja furioso o tempo todo. Se não houver oponente para Tauros lutar, ele atacará árvores grossas e as derrubará para se acalmar."),
-    MAGIKARP("Pokémon Peixe.\n\nMagikarp é uma desculpa patética para um Pokémon que só é capaz de fracassar e espirrar água. Esse comportamento levou os cientistas a realizar pesquisas sobre ele."),
-    GYARADOS("Pokémon Atrocidade.\n\nQuando Magikarp evolui para Gyarados, suas células cerebrais passam por uma transformação estrutural. Dizem que essa transformação é responsável pela natureza violenta desse Pokémon."),
-    LAPRAS("Pokémon Transporte.\n\nAs pessoas levaram Lapras quase ao ponto de extinção. À noite, diz-se que este Pokémon canta lamentosamente enquanto procura os poucos que ainda restam."),
-    DITTO("Pokémon Transformação.\n\nDitto reorganiza sua estrutura celular para se transformar em outras formas. No entanto, se tentar se transformar em algo, confiando em sua memória, este Pokémon consegue obter detalhes errados."),
-    EEVEE("Pokémon Evolução.\n\nEevee tem uma composição genética instável que de repente sofre mutações devido ao ambiente em que vive. A radiação de várias pedras faz com que este Pokémon evolua."),
-    VAPOREON("Pokémon Jato de Bolhas.\n\nVaporeon sofreu uma mutação espontânea e desenvolveu barbatanas e brânquias que lhe permitem viver debaixo d'água. Este Pokémon tem a capacidade de controlar livremente a água."),
-    JOLTEON("Pokémon Relâmpago.\n\nAs células de Jolteon geram um baixo nível de eletricidade. Esse poder é amplificado pela eletricidade estática de seu pêlo, permitindo que o Pokémon solte raios. O pêlo eriçado é feito de agulhas carregadas eletricamente."),
-    FLAREON("Pokémon Chama.\n\nO pêlo macio de Flareon tem um objetivo funcional - libera calor no ar para que seu corpo não fique excessivamente quente. A temperatura corporal deste Pokémon pode subir para um máximo de 900°C."),
-    PORYGON("Pokémon Virtual.\n\nPorygon é capaz de se reverter inteiramente em dados e entrar no ciberespaço. Este Pokémon é protegido contra cópia, portanto não pode ser duplicado por cópia."),
-    OMANYTE("Pokémon Espiral.\n\nOmanyte é um dos Pokémon antigos e já extintos que foram regenerados a partir de fósseis por pessoas. Se atacado por um inimigo, ele se esconde dentro de sua casca dura."),
-    OMASTAR("Pokémon Espiral.\n\nOmastar usa seus tentáculos para capturar suas presas. Acredita-se que ele tenha sido extinto porque sua concha ficou muito grande e pesada, fazendo com que seus movimentos se tornassem muito lentos e pesados."),
-    KABUTO("Pokémon Crustáceo.\n\nKabuto é um Pokémon que foi regenerado a partir de um fóssil. No entanto, em casos extremamente raros, exemplos vivos foram descobertos. O Pokémon não mudou em nada por 300 milhões de anos."),
-    KABUTOPS("Pokémon Crustáceo.\n\nKabutops nadavam debaixo d'água para caçar suas presas nos tempos antigos. Aparentemente, o Pokémon estava evoluindo de habitante da água para viver em terra, como é evidente desde o início da mudança em suas brânquias e pernas."),
-    AERODACTYL("Pokémon Fóssil.\n\nAerodactyl é um Pokémon da era dos dinossauros. Foi regenerado a partir de material genético extraído de âmbar. Imagina-se que tenha sido o rei dos céus nos tempos antigos."),
-    SNORLAX("Pokémon Dormindo.\n\nO dia típico de Snorlax consiste em nada mais do que comer e dormir. É um Pokémon tão dócil que há crianças que usam sua barriga enorme como um lugar para brincar."),
-    ARTICUNO("Pokémon Geada.\n\nArticuno é um Pokémon Pássaro Lendário, que pode controlar o gelo. O bater de suas asas esfria o ar. Como resultado, diz-se que quando este Pokémon voa, a neve cai."),
-    ZAPDOS("Pokémon Elétrico.\n\nZapdos é um Pokémon Pássaro Lendário, que tem a capacidade de controlar a eletricidade. Geralmente vive em nuvens de trovoada. O Pokémon ganha poder se for atingido por raios."),
-    MOLTRES("Pokémon Chama.\n\nMoltres é um Pokémon Pássaro Lendário, que tem a capacidade de controlar o fogo. Se este Pokémon for ferido, é dito que ele mergulha seu corpo no magma derretido de um vulcão para queimar e curar a si próprio."),
-    DRATINI("Pokémon Dragão.\n\nDratini muda constantemente e descama sua pele velha. Isso ocorre porque a energia da vida dentro de seu corpo aumenta constantemente para atingir níveis incontroláveis."),
-    DRAGONAIR("Pokémon Dragão.\n\nDragonair armazena uma enorme quantidade de energia dentro de seu corpo. Diz-se que altera as condições climáticas nas proximidades, descarregando energia dos cristais no pescoço e na cauda."),
-    DRAGONITE("Pokémon Dragão.\n\nDragonite é capaz de dar a volta no globo em apenas 16 horas. É um Pokémon de bom coração que leva navios perdidos e afundando numa tempestade para a segurança da terra."),
-    MEWTWO("Pokémon Genético.\n\nMewtwo é um Pokémon que foi criado por manipulação genética. No entanto, embora o poder científico dos humanos tenha criado o corpo desse Pokémon, eles não conseguiram dotar Mewtwo de um coração compassivo."),
-    MEW("Pokémon Novas Espécies.\n\nDiz-se que Mew possui a composição genética de todos os Pokémon. Ele é capaz de se tornar invisível à vontade, por isso evita completamente a percepção, mesmo que se aproxime das pessoas."),
-
-    //Segunda geração
-
-    CHIKORITA("Pokémon Folha.\n\nNa batalha, Chikorita balança sua folha para manter o inimigo à distância. No entanto, uma fragrância doce também flutua da folha, acalmando os Pokémon na batalha e criando uma atmosfera acolhedora e amigável ao redor."),
-    BAYLEEF("Pokémon Folha.\n\nO pescoço de Bayleef é rodeado por folhas enroladas. Dentro de cada folha tubular há um pequeno rebento de uma árvore. A fragrância destas mudas deixa as pessoas animadas."),
-    MEGANIUM("Pokémon Erva.\n\nA fragrância da flor de Meganium abranda e acalma emoções. Na batalha, este Pokémon exala mais de seu perfume calmante para atenuar o espírito de luta do inimigo."),
-    CYNDAQUIL("Pokémon Rato de Fogo.\n\nCyndaquil se protege acendendo as chamas nas costas. As chamas são vigorosas se o Pokémon estiver com raiva. No entanto, se estiver cansado, as chamas crepitam com combustão incompleta."),
-    QUILAVA("Pokémon Vulcão.\n\nQuilava mantém seus inimigos afastados com a intensidade de suas chamas e rajadas de ar superaquecido. Este Pokémon aplica sua extraordinária agilidade para evitar ataques, mesmo enquanto queima o inimigo com chamas."),
-    TYPHLOSION("Pokémon Vulcão.\n\nTyphlosion se oculta por trás de uma névoa de calor cintilante criada usando suas chamas intensamente quentes. Este Pokémon cria explosões ardentes que queimam tudo até sobrarem cinzas."),
-    TOTODILE("Pokémon Grande Mandíbula.\n\nApesar da pequenez do corpo, as mandíbulas de Totodile são muito poderosas. Enquanto o Pokémon pode pensar que é apenas um beliscão de brincadeira, sua mordida tem poder suficiente para causar ferimentos graves."),
-    CROCONAW("Pokémon Grande Mandíbula.\n\nUma vez que Croconaw aperta as mandíbulas no inimigo, ele absolutamente não o solta. Como as pontas de suas presas são bifurcadas como anzóis farpados, tornam-se impossíveis de remover quando se afundam."),
-    FERALIGATR("Pokémon Grande Mandíbula.\n\nFeraligatr intimida seus inimigos abrindo sua boca enorme. Na batalha, ele chutará o chão com força, com suas patas traseiras grossas e poderosas para atacar o inimigo a uma velocidade incrível."),
-    SENTRET("Pokémon Sentinela.\n\nQuando Sentret dorme, o faz enquanto outro fica de guarda. A sentinela acorda os outros ao primeiro sinal de perigo. Quando este Pokémon se separa de sua matilha, torna-se incapaz de dormir devido ao medo."),
-    FURRET("Pokémon Corpo Longo.\n\nFurret tem uma construção muito fina. Quando está sob ataque, ele pode se contorcer com facilidade em espaços estreitos e fugir. Apesar de seus membros curtos, este Pokémon é muito ágil e ligeiro."),
-    HOOTHOOT("Pokémon Coruja.\n\nTodos os dias, inclina a cabeça no mesmo ritmo. Há muito tempo, as pessoas criaram esses Pokémon para servir como relógios."),
-    NOCTOWL("Pokémon Coruja.\n\nCom olhos que podem ver na escuridão, nunca deixa sua presa escapar. Alguns até chamam de \"o imperador das noites escuras\"."),
-    LEDYBA("Pokémon Cinco Estrelas.\n\nÉ muito tímido. Ele terá medo de se mover se estiver sozinho. Mas estará ativo se estiver em um grupo. Quando o tempo fica frio, muitos Ledyba se reúnem de todos os lugares para se agrupar e manter um ao outro quente."),
-    LEDIAN("Pokémon Cinco Estrelas.\n\nDizem que os padrões nas costas estão relacionados às estrelas no céu noturno, mas os detalhes dessa relação permanecem obscuros."),
-    SPINARAK("Pokémon Cospe-teia.\n\nCom fios da boca, tece teias robustas que não quebram, mesmo que você as atinja."),
-    ARIADOS("Pokémon Patas Longas.\n\nToda noite, vagueia em busca de presas, cujos movimentos restringe ao soltar teias antes de os morder com suas mandíbulas."),
-    CROBAT("Pokémon Morcego.\n\nSuas presas são tão afiadas que se ele te morde no escuro e suga seu sangue, você não notará nenhuma dor ou perceberá que foi mordido."),
-    CHINCHOU("Pokémon Pescador.\n\nSuas duas antenas brilham suavemente para atrair as presas, tornando-o um Pokémon útil para a pesca noturna."),
-    LANTURN("Pokémon Luz.\n\nQuando as bactérias que vivem dentro de suas antenas absorvem os fluidos corporais de Lanturn, um forte efeito luminescente é produzido."),
-    PICHU("Pokémon Rato Pequeno.\n\nAinda não consegue usar bem a eletricidade. Quando é surpreendido ou animado, descarrega eletricidade sem querer"),
-    CLEFFA("Pokémon Forma de Estrela.\n\nNas noites com muitas estrelas cadentes, eles se reúnem em bandos e dançam em círculos. Se você os vir, algo de bom acontecerá!"),
-    IGGLYBUFF("Pokémon Balão.\n\nEstá sempre praticando seu canto porque quer melhorar. Mesmo dormindo, ele continua cantando nos seus sonhos!"),
-    TOGEPI("Pokémon Bola Pontuda.\n\nComo energia, Togepi usa as emoções positivas de compaixão e prazer emanadas por pessoas e Pokémon. Este Pokémon armazena sentimentos de felicidade dentro de sua concha, depois os compartilha com os outros."),
-    TOGETIC("Pokémon Felicidade.\n\nDizem que Togetic é um Pokémon que traz boa sorte. Quando o Pokémon vê alguém que é puro de coração, diz-se que ele aparece e compartilha sua felicidade com essa pessoa."),
-    NATU("Pokémon Passarinho.\n\nEmbora ainda não consiga voar, seu poder de saltar é notável. Ele pula para as árvores e arranca os brotos dos galhos."),
-    XATU("Pokémon Místico.\n\nEmbora possa ver o futuro, não deseja alterá-lo, e é provavelmente por isso que permanece imóvel o tempo todo."),
-    MAREEP("Pokémon Lã.\n\nEsfregar o seu velo gera eletricidade. Você vai querer acariciá-lo porque é fofo, mas se você usar a mão nua, receberá um choque doloroso."),
-    FLAAFFY("Pokémon Lã.\n\nNos locais onde o velo não cresce, a pele é emborrachada e não conduz eletricidade. Esses pontos são seguros para tocar."),
-    AMPHAROS("Pokémon Luz.\n\nA luz da cauda pode ser vista do espaço. É por isso que você sempre pode dizer exatamente onde está, e é por isso que geralmente mantém a luz apagada."),
-    BELLOSSOM("Pokémon Flor.\n\nQuando Bellossom fica exposto a muita luz do sol, as folhas que circundam seu corpo começam a girar. A dança deste Pokémon é conhecida nas terras do sul."),
-    MARILL("Pokémon Rato Aquático.\n\nA cauda cheia de óleo de Marill age como um colete salva-vidas. Se você vir apenas a cauda balançando na superfície da água, é uma indicação certa de que este Pokémon está mergulhando sob a água para se alimentar de plantas aquáticas."),
-    AZUMARILL("Pokémon Coelho Aquático.\n\nAs orelhas compridas de Azumarill são sensores indispensáveis. Ao focar sua audição, este Pokémon pode identificar quais tipos de presas estão em volta, mesmo em rios agitados e rápidos."),
-    SUDOWOODO("Pokémon Imitação.\n\nÉ tão popular entre os idosos que há uma revista dedicada a este Pokémon. Os fãs ficam obcecados com o comprimento e o ângulo específicos de seus braços."),
-    POLITOED("Pokémon Sapo.\n\nQuanto mais longos e luxuosos forem os cabelos encaracolados na cabeça, maior será o número de Poliwag e Poliwhirl que os obedecerão."),
-    HOPPIP("Pokémon Erva de Algodão.\n\nEste Pokémon flutua e desliza com o vento. Se sentir a aproximação dos ventos fortes, o Hoppip liga suas folhas com outros Hoppip para se preparar para não ser surpreendido."),
-    SKIPLOOM("Pokémon Erva de Algodão.\n\nA flor do Skiploom floresce quando a temperatura sobe acima de 16 graus Celsius. Quanto a flor abre depende da temperatura. Por esse motivo, esse Pokémon às vezes é usado como termômetro."),
-    JUMPLUFF("Pokémon Erva de Algodão.\n\nJumpluff usa os ventos quentes do sul para atravessar o mar e voar para terras estrangeiras. O Pokémon desce ao chão quando encontra ar frio enquanto está flutuando."),
-    AIPOM("Pokémon Cauda Longa.\n\nEnquanto usava cada vez mais o rabo, suas mãos se tornaram desajeitadas. Faz o seu ninho no alto das copas das árvores."),
-    SUNKERN("Pokémon Semente.\n\nSunkern tenta se mover o mínimo possível. Faz isso porque tenta conservar todos os nutrientes que armazenou em seu corpo para sua evolução. Não comerá nada, subsistindo apenas no orvalho da manhã."),
-    SUNFLORA("Pokémon Sol.\n\nSunflora converte energia solar em nutrição. Ele se move ativamente durante o dia, quando está quente. Ele para de se mover assim que o sol se põe."),
-    YANMA("Pokémon Asas Claras.\n\nYanma é capaz de ver 360 graus sem ter que mover os olhos. É um grande aviador, capaz de fazer paradas bruscas e virar no ar. Este Pokémon usa sua habilidade de voar para perseguir rapidamente presas avistadas."),
-    WOOPER("Pokémon Peixe de Água.\n\nWooper geralmente vive na água. No entanto, ocasionalmente sai à terra em busca de comida. Em terra, ele reveste o corpo com uma camada tóxica e pegajosa."),
-    QUAGSIRE("Pokémon Peixe de Água.\n\nQuagsire procura comida, deixando a boca aberta na água e esperando a presa desavisada. Como o Pokémon não se move, não fica com muita fome."),
-    ESPEON("Pokémon Sol.\n\nO poder psíquico se acumula na esfera em sua testa enquanto se banha ao sol. Espeon não é bom em lutar à noite."),
-    UMBREON("Pokémon Luar.\n\nEspreita no escuro da noite procurando presas. No momento em que ele ataca, os anéis em seu corpo brilham fracamente, mas de maneira ameaçadora."),
-    MURKROW("Pokémon Escuridão.\n\nEle procura coisas brilhantes para seu chefe. Dizem que a presença de Murkrow traz má sorte, muitas pessoas a detestam."),
-    SLOWKING("Pokémon Real.\n\nEstá constantemente surgindo com novas idéias que mudariam o mundo, mas assim que se depara com uma nova idéia, ela a esquece."),
-    MISDREAVUS("Pokémon Guincho.\n\nO que dá sentido à sua vida é assustar os outros. Se você colocar o ouvido contra as esferas vermelhas em volta do pescoço, poderá ouvir um grito agudo."),
-    UNOWN("Pokémon Símbolo.\n\nEste Pokémon tem a forma de escritas antigas. É um mistério sobre o que veio primeiro, os escritos antigos ou os vários Unown. A pesquisa sobre este tópico está em andamento, mas nada se sabe."),
-    WOBBUFFET("Pokémon Paciente.\n\nWobbuffet não faz nada além de suportar ataques - não ataca por si próprio. No entanto, ele não aguenta um ataque na cauda. Quando isso acontecer, o Pokémon tentará levar o inimigo usando Vínculo do Destino."),
-    GIRAFARIG("Pokémon Pescoço Longo.\n\nA cabeça traseira de Girafarig também tem um cérebro, mas é pequeno. A cabeça traseira ataca em resposta a cheiros e sons. Aproximar-se deste Pokémon por trás pode fazer com que a cabeça traseira repentinamente ataque e morda."),
-    PINECO("Pokémon Mariposa.\n\nCola cascas de árvores com saliva em si mesmo, tornando-se mais espesso e maior. Pineco idosos são ridiculamente enormes."),
-    FORRETRESS("Pokémon Mariposa.\n\nQuando algo se aproxima, ele dispara fragmentos de sua concha de aço em ataque. Esta não é uma ação consciente, mas um reflexo condicionado."),
-    DUNSPARCE("Pokémon Serpente Longa.\n\nQuando vê uma pessoa, cava um buraco com a cauda para escapar. Se você encontrar um, considere-se com sorte."),
-    GLIGAR("Pokémon Escorpião Voador.\n\nGligar plana pelo ar sem emitir nenhum som, como se estivesse deslizando. Este Pokémon se agarra à face de seu inimigo usando suas patas traseiras com garras e as pinças grandes em suas pernas dianteiras, depois injeta a presa com sua farpa venenosa."),
-    STEELIX("Pokémon Serpente de Ferro.\n\nSteelix vive ainda mais subterrâneo que Onix. Este Pokémon é conhecido por cavar em direção ao núcleo da Terra. Existem registros deste Pokémon atingindo uma profundidade de mais de 100 quilômetros."),
-    SNUBBULL("Pokémon Fada.\n\nAo contrário de sua aparência, é bastante tímido. Ao brincar com outros Pokémon filhotes, às vezes é maltratado."),
-    GRANBULL("Pokémon Fada.\n\nEmbora tenha mandíbulas poderosas, não se importa com disputas, portanto raramente tem a chance de mostrar sua força."),
-    QWILFISH("Pokémon Balão.\n\nQwilfish suga a água, inflando-se. Este Pokémon usa a pressão da água que ingeriu para atirar setas tóxicas de uma só vez de todo o corpo. Acha a natação um tanto desafiadora."),
-    SCIZOR("Pokémon Pinça.\n\nUma vez que identificar algo como inimigo, continuará batendo nele com suas duras pinças de aço até que não restem nada além de restos."),
-    SHUCKLE("Pokémon Fungo.\n\nShuckle se esconde silenciosamente debaixo de pedras, mantendo seu corpo escondido dentro de sua casca dura enquanto come frutas que guardou. As bagas se misturam com seus fluidos corporais para se tornar um suco."),
-    HERACROSS("Pokémon Unicórnio.\n\nPercorre florestas procurando néctar doce. Embora possua uma força física fantástica, não é tão bom em voar."),
-    SNEASEL("Pokémon Garras Afiadas.\n\nEles vão cooperar para roubar ovos dos ninhos dos Pokémon pássaros, mas brigam para determinar qual deles comerá os ovos."),
-    TEDDIURSA("Pokémon Ursinho.\n\nEste Pokémon gosta de lamber as palmas das mãos, adoçadas por estarem embebidas em mel. Teddiursa inventa seu próprio mel misturando frutas e pólen coletados por Beedrill."),
-    URSARING("Pokémon Hibernador.\n\nNas florestas habitadas por Ursaring, diz-se que existem muitos riachos e árvores altas onde eles colhem comida. Este Pokémon percorre sua floresta coletando comida todos os dias."),
-    SLUGMA("Pokémon Lava.\n\nO magma derretido percorre todo o sistema circulatório de Slugma. Se esse Pokémon é resfriado, o magma esfria e endurece. Seu corpo fica quebradiço e os pedaços caem, reduzindo seu tamanho."),
-    MAGCARGO("Pokémon Lava.\n\nA concha de Magcargo é na verdade sua pele que endureceu como resultado do resfriamento. Sua concha é muito quebradiça e frágil - apenas tocá-la faz com que ela se desfaça. Este Pokémon retorna ao seu tamanho original mergulhando-se em magma."),
-    SWINUB("Pokémon Porco.\n\nSwinub procura comida esfregando o focinho no chão. Sua comida favorita é um cogumelo que cresce sob a cobertura de grama morta. Este Pokémon ocasionalmente procura fontes termais."),
-    PILOSWINE("Pokémon Suíno.\n\nPiloswine é coberto por uma espessa camada de pelo comprido que permite suportar o frio congelante. Este Pokémon usa suas presas para desenterrar alimentos que foram enterrados no gelo."),
-    CORSOLA("Pokémon Coral.\n\nAs pontas na cabeça absorvem os nutrientes da água limpa. Eles não podem sobreviver na água poluída."),
-    REMORAID("Pokémon Jato.\n\nEste Pokémon se liga a Mantine e compartilha sua prosperidade. Quando seu Mantine é atacado, Remoraid luta ao lado dele!"),
-    OCTILLERY("Pokémon Jato.\n\nA tinta que cospe ao escapar é especial. Ela contém uma substância que entorpece o sentido do olfato, para que os Pokémon com nariz aguçado se percam."),
-    DELIBIRD("Pokémon Entrega.\n\nEnvolve comida na cauda para carregá-la. O Delibird com a cauda maior é o chefe do rebanho!"),
-    MANTINE("Pokémon Pipa.\n\nMuitas pessoas começaram a surfar porque admiram o quão magnificamente Mantine nada."),
-    SKARMORY("Pokémon Pássaro com Armadura.\n\nSeu corpo está envolto em armadura de aço. Parece pesado, mas pode voar a velocidades de até 185 milhas por hora!"),
-    HOUNDOUR("Pokémon Escuridão.\n\nEle coopera com os outros com habilidade. Quando se torna seu parceiro, é muito leal a você como treinador e obedecerá às suas ordens."),
-    HOUNDOOM("Pokémon Escuridão.\n\nEles lançam chamas misturadas com veneno para acabar com seus oponentes. Eles dividem suas presas igualmente entre os membros de sua matilha."),
-    KINGDRA("Pokémon Dragão.\n\nKingdra vive em profundidades extremas do oceano que, de outra forma, são desabitadas. Acredita-se há muito tempo que o bocejo deste Pokémon cria correntes oceânicas em espiral."),
-    PHANPY("Pokémon Nariz Longo.\n\nComo seu ninho, Phanpy cava um poço vertical no chão à beira de um rio. Ele marca a área ao redor do ninho com seu tronco para que os outros saibam que a área foi reivindicada."),
-    DONPHAN("Pokémon Armadura.\n\nO ataque favorito de Donphan é enrolar seu corpo em uma bola e depois atacar seu inimigo enquanto rola em alta velocidade. Uma vez que começa a rolar, este Pokémon não pode parar com muita facilidade."),
-    PORYGON2("Pokémon Virtual.\n\nUma inteligência artificial foi instalada nele. Aprende várias coisas por conta própria, mas lembra até de coisas que não precisa saber."),
-    STANTLER("Pokémon Grande Chifre.\n\nOs magníficos chifres de Stantler eram comercializados a preços altos como obras de arte. Como resultado, este Pokémon foi caçado e quase extinto por aqueles que estavam atrás dos chifres de valor inestimável."),
-    SMEARGLE("Pokémon Pintor.\n\nEle desenha símbolos com o fluido que escorre da ponta da cauda. Dependendo do símbolo, os fanáticos por Smeargle pagarão muito dinheiro por eles."),
-    TYROGUE("Pokémon Briga.\n\nTyrogue fica estressado se não conseguir treinar todos os dias. Ao criar este Pokémon, o treinador deve estabelecer e encorajar vários métodos de treinamento."),
-    HITMONTOP("Pokémon Parada de Mão.\n\nO Hitmontop gira de cabeça pra baixo em alta velocidade, sempre dando chutes. Essa técnica é uma mistura notável de ataque e defesa ao mesmo tempo. O Pokémon viaja mais rápido girando do que andando."),
-    SMOOCHUM("Pokémon Beijo.\n\nQuando examina as coisas, antes de tocá-las com as mãos, faz contato com os lábios e depois lambe cuidadosamente toda a sujeira dos lábios."),
-    ELEKID("Pokémon Elétrico.\n\nQuando ouve o estrondo de trovão, o humor de Elekid melhora. Pode ser útil gravar esse som e tocá-lo quando Elekid estiver triste."),
-    MAGBY("Pokémon Carvão Vivo.\n\nSua temperatura corporal é sempre em torno de 600 graus Celsius. Se Magby cair em um lago pequeno o suficiente, tudo ficará seco."),
-    MILTANK("Pokémon Vaca Leiteira.\n\nProduz mais de cinco litros de leite por dia. Quanto maior a qualidade das pastagens em que vive, mais rico e saboroso fica o leite."),
-    BLISSEY("Pokémon Felicidade.\n\nSeus ovos são tão delicados que alguns dizem que comer um lhe trará felicidade. Esses ovos alcançam os preços mais altos do mercado."),
-    RAIKOU("Pokémon Trovão.\n\nRaikou encarna a velocidade do raio. Os rugidos deste Pokémon emitem ondas de choque tremendo no ar e sacodem o chão como se relâmpagos estivessem caindo."),
-    ENTEI("Pokémon Vulcão.\n\nEntei encarna a paixão do magma. Pensa-se que este Pokémon nasceu na erupção de um vulcão. Emite enormes explosões de fogo que consomem totalmente tudo o que tocam."),
-    SUICUNE("Pokémon Aurora.\n\nSuicune encarna a compaixão de uma pura fonte de água. Atravessa a terra com graça. Este Pokémon tem o poder de purificar a água poluída."),
-    LARVITAR("Pokémon Pele de Pedra.\n\nNascido no subsolo, ele passa pela terra até a superfície, onde estão seus pais. Não lida bem com a luz brilhante que encontra lá."),
-    PUPITAR("Pokémon Casca Dura.\n\nEste Pokémon problemático mal pode esperar para evoluir, então alivia o estresse se movendo descontroladamente usando gás comprimido."),
-    TYRANITAR("Pokémon Armadura.\n\nEste Pokémon é um desastre móvel, deixando montanhas desmoronadas e casas destruídas em seu rastro."),
-    LUGIA("Pokémon Mergulho.\n\nAs asas de Lugia têm um poder devastador - uma leve vibração de suas asas pode explodir casas comuns. Como resultado, este Pokémon escolhe viver fora da vista no fundo do mar."),
-    HOOH("Pokémon Arco-íris.\n\nAs penas de Ho-Oh brilham em sete cores, dependendo do ângulo em que são atingidas pela luz. Diz-se que essas penas trazem felicidade aos portadores. Diz-se que este Pokémon vive ao pé de um arco-íris."),
-    CELEBI("Pokémon Viagem no Tempo.\n\nEste Pokémon veio do futuro cruzando ao longo do tempo. Pensa-se que enquanto Celebi aparecer, um futuro claro e brilhante nos espera.")
-}
-//NOME("Pokémon .\n\n"),
