--- app/src/main/java/com/example/pokedex/view/SplashActivity.kt
@@ -1,25 +1,0 @@
-package com.example.pokedex.view
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.os.Handler
-import com.example.pokedex.R
-
-class SplashActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_splash)
-
-        //Delay pra testar a tela inicial
-        Handler().postDelayed({ iniciaMain() }, 500)
-    }
-
-    //Inicia a MainActivity
-    fun iniciaMain() {
-        val intent = Intent(this, MainActivity::class.java)
-        startActivity(intent)
-        finish()
-    }
-}
